Basic readme and stable package - building tests with package
