+++ clients/node-client/src/index.ts
@@ -1,11 +1,27 @@
 export { Chunkr } from "./Chunkr";
 export { TaskResponse } from "./models/TaskResponse";
-export { Configuration, Status, OcrStrategy } from "./models/Configuration";
+export {
+  Configuration,
+  JsonSchema,
+  Property,
+  OcrStrategy,
+  ChunkProcessing,
+  SegmentProcessing,
+  SegmentationStrategy,
+  GenerationStrategy,
+  GenerationConfig,
+  CroppingStrategy,
+} from "./models/Configuration";
 export type {
   TaskResponseData,
+  Status,
   Output,
   Segment,
   Chunk,
   BoundingBox,
+  OCRResult,
+  SegmentType,
+  ExtractedField,
+  ExtractedJson,
 } from "./models/TaskResponseData";
 export type { ClientConfig } from "./models/ClientConfig";
